GPGPU: CL: Do not assign to an array from an initializer list
Arrays can only be initialized when declared, but an existing array
cannot be assigned from an initializer list.

For example:
    size_t data.local[] = {1, 2, 3};
is valid, and compiles okay.

On the other hand:
    data.local = {1, 2, 3};
is invalid, and depending on the compiler version may or may not
compile. To fix this, assign each member of the array individually.