GPGPU: OpenCL: Drop dynamic loading of OpenCL library
When the loader in OpenCL_Dyn_Load.c was introduced, OpenCL support
was new, and usually required the installation  of bloated binary only
packages. Then it made sense to make OpenCL an optional dependency.

With the introduction of the installable client drivers, the OpenCL
library became a shim to link to each vendor's OpenCL implementation.
Thus it is possible today to link against OpenCL without needing an
actual OpenCL implementation to be installed -- exactly what
OpenCL_Dyn_Load.c provided.

It is no longer justifiable to carry a custom OpenCL loader, so make
OpenCL a mandatory dependency, and drop OpenCL_Dyn_Load.c