GPGPU: OpenCL: Do not use clGetDeviceInfo() to get platform ID
On mesa OpenCL (Clover), clGetDeviceInfo() will return NULL for
CL_DEVICE_PLATFORM. This value is used to populate the
clDeviceProp::platform member, later used in clCreateContext(). When
this is NULL, we can't create an OpenCL context and everything else
fails.

We already know the platform ID in the clPlatformProp constructor. We
can avoid calling clGetDeviceInfo(CL_DEVICE_PLATFORM) by manually
populating the clDeviceProp::platform field. This was tested to work
with both clover and AMD APP.