ElectroMag: Rework simulation parameter passing
The old way of selecting parameters involved a different commandline
flag for each simulation size. It was messy because adding or removing
parameters required synchronized changes in four different places.

Replace the old system with a system where each set of parameters has
a human-readable name. The name is passed as the '--simsize" argument.
An invalid "--simsize" value prints the available levels, and exits.s