GPGPU_Segment: CL_Manager: Add method to select platform by name
In order to create an OpenCL context, one needs to know the platform
and device IDs. They are available via ClManager:platforms, which  is
private. The only way to access this member was via
ClManager::fstGetPlats(), which spits out all the platforms and lets
the user figure it all out.

We prefer to provide an intuitive way of getting the desired platform.
One such method is to use a partial platform name, as the key, and
return the first matching platform. Although this won't guarantee
uniqueness, it's not common to have similarly named OpenCL platforms.
Thus it's very hard to have non-unique keys, unless they are a
character or two long.

One advantage of this approach is that an empty platform name, will
match the first available platform. Implement this feature under the
FindPlatformByName() name.