ElectroMag: Select CL platform with "--clplatform" parameter
It is not uncommon today to have several OpenCL providers on the same
system. It's even possible to have different providers for the same
compute device. It is no longer sufficient to pick the first available
platform and crunch numbers.

Add a "--clplatform" parameter which takes a platform name as an
argument. This is passed to the functor via TestCL(). In the event
"--clplatform" is not given, the behavior is still to pick the first
available platform.