GPGPU_Segment: CL_Manager.hpp: Remove excessive OpenCL comments
The members of ClManager::clDeviceProp are 1:1 equvalent to OpenCL
CL_DEVICE_* properties. They appear to be copypasted from the OpenCL
documentation. It makes sense to know which OpenCL property each
member maps to, but beyond that, users should refer to the OpenCL
documentation.

We don't try to duplicate the OpenCL documentation, or provide OpenCL
in the header file. And whenever the official documentation is
updated, we're out of sync. Shorten the member comments to only
contain the CL_DEVICE_* property that the member maps to.