GPGPU_Segment: CL_Electrostatics: Remove dead code
Leaving dead code around, whether it's commented out or #if 0'd, is
akin to leaving one's dirty underwear laying around in the open. And
when it sits for over a decade, the results can be quite aromatic.

Remove code that is commented out. It's not used, so it's obviously
not needed.