GPGPU_Segment: Electrostatics.cpp: Remove CalcField()
CalcField() used to be the entry point for GPU computation. Since the
move to OpenCL the entry point is now TestCL(). CalcField() doesn't
have an interface class, or even a header definition. It's just a
vestigial appendage. Remove it.

There needs to be an interface class defining entry points and
parameters for static field calculations. The interface should be
shared by both CPU and OpenCL/GPU code. Such an interface is beyond
the scope of this change.